Extract the credits meter; use imported icons in the footer
Move the free-credits meter into its own NavFooterCreditsRow component with
co-located CSS, and let it own the NavFooterCredits contract (the two
useFreeCreditsSummary seams now import the type from there). NavFooter is left
as the composition: build the row list, separate with dividers, render.

Swap the two hand-rolled SVGs for the icons the rest of the app imports -
@mui/icons-material Settings on the account row (what the editor footer used
before this branch) and Link in the story's link-account stand-in.
