+++ frontend/editor/src/core/components/shared/navFooter/NavFooter.css
@@ -87,56 +87,6 @@
   font-size: 0.8125rem;
 }
 
-/* ---- Credits meter ---- */
-
-.sui-nav-footer__credits {
-  flex-direction: column;
-  align-items: stretch;
-  gap: 0.375rem;
-  cursor: default;
-}
-
-.sui-nav-footer__credits:not(:disabled):hover {
-  background: none;
-}
-
-.sui-nav-footer__credits-head {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 0.8125rem;
-}
-
-.sui-nav-footer__dot {
-  width: 0.4375rem;
-  height: 0.4375rem;
-  border-radius: 50%;
-  flex-shrink: 0;
-  background-color: var(--c-success);
-}
-.sui-nav-footer__dot[data-tone="warning"] {
-  background-color: var(--c-warning);
-}
-.sui-nav-footer__dot[data-tone="danger"] {
-  background-color: var(--c-danger);
-}
-
-.sui-nav-footer__credits-label {
-  flex: 1;
-  min-width: 0;
-  font-weight: 500;
-  color: var(--c-text);
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.sui-nav-footer__credits-count {
-  flex-shrink: 0;
-  color: var(--c-text-muted);
-  font-variant-numeric: tabular-nums;
-}
-
 /* ---- Collapsed icon rail ---- */
 
 .sui-nav-footer[data-collapsed] {
