+++ frontend/editor/src/core/components/shared/navFooter/NavFooterCreditsRow.css
@@ -1,0 +1,51 @@
+/* Free-credits meter inside the sidebar footer. The row base (padding, hover,
+   focus) comes from NavFooter.css; these rules are the meter itself. */
+
+.sui-nav-footer__credits {
+  flex-direction: column;
+  align-items: stretch;
+  gap: 0.375rem;
+  cursor: default;
+}
+
+/* Not actionable, so it must not read as hoverable. */
+.sui-nav-footer__credits:not(:disabled):hover {
+  background: none;
+}
+
+.sui-nav-footer__credits-head {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  font-size: 0.8125rem;
+}
+
+.sui-nav-footer__dot {
+  width: 0.4375rem;
+  height: 0.4375rem;
+  border-radius: 50%;
+  flex-shrink: 0;
+  background-color: var(--c-success);
+}
+.sui-nav-footer__dot[data-tone="warning"] {
+  background-color: var(--c-warning);
+}
+.sui-nav-footer__dot[data-tone="danger"] {
+  background-color: var(--c-danger);
+}
+
+.sui-nav-footer__credits-label {
+  flex: 1;
+  min-width: 0;
+  font-weight: 500;
+  color: var(--c-text);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.sui-nav-footer__credits-count {
+  flex-shrink: 0;
+  color: var(--c-text-muted);
+  font-variant-numeric: tabular-nums;
+}
